pyproject.toml
[tool.poetry]
name = "pyhex"
version = "0.1.0"
description = ""
authors = ["Alexander Dadiani"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
alembic = "^1.7.5"
SQLAlchemy = "^1.4.27"
pydantic = "^1.8.2"
python-dotenv = "^0.19.2"
psycopg2 = "^2.9.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
mypy = "^0.910"
flake8 = "^4.0.1"
isort = "^5.10.1"
factory-boy = "^3.2.1"
pytest-cov = "^3.0.0"
ipdb = "^0.13.9"
requests = "^2.27.1"
pre-commit = "^2.17.0"

[tool.mypy]
ignore_missing_imports = true
follow_imports = "silent"
follow_imports_for_stubs = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_configs = true
no_implicit_optional = true
warn_unreachable = true
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
